Finally fixed most audio issues

+++ src/lib.rs
@@ -19,6 +19,6 @@ pub trait FrameSink {
 }
 
 pub trait AudioSink {
-    fn push_sample(&mut self, sample: (i16, i16)) -> bool;
-    fn push_samples(&mut self, samples: &mut VecDeque<i16>);
+    fn push_sample(&mut self, sample: (f32, f32)) -> bool;
+    fn push_samples(&mut self, samples: &mut VecDeque<f32>);
 }
